pyproject.toml
[tool.poetry]
name = "counter"
version = "0.1.0"
description = ""
authors = ["Giovanni Go <giovanni@jamcity.com>"]
readme = "README.md"
packages = [{include = "counter_aiohttp"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.4"
aioredis = "^2.0.1"
gunicorn = "^20.1.0"
daphne = "^4.0.0"
aiohttp = {extras = ["speedups"], version = "^3.8.3"}
locust = "^2.14.0"
mysqlclient = "^2.1.1"
uvloop = "^0.17.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
